[Filters] Clean up passing date

+++ app/src/main/java/com/company/activityart/presentation/edit_art_screen/EditArtModels.kt
@@ -13,8 +13,8 @@ import kotlinx.parcelize.RawValue
 
 sealed interface EditArtViewEvent : ViewEvent {
     data class FilterDateChanged(
-        val newUnixSecondStart: Float,
-        val newUnixSecondEnd: Float
+        val newUnixSecondStart: Long,
+        val newUnixSecondEnd: Long
     ) : EditArtViewEvent
 
     sealed interface FilterTypeChanged : EditArtViewEvent {
